+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.60.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.60.0...@kong-ui-public/entities-shared@3.60.1) (2026-08-06)
+
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#3662](https://github.com/Kong/public-ui-components/issues/3662)) ([6ae03ce](https://github.com/Kong/public-ui-components/commit/6ae03cea16634445849a1b9dd693bb40c86be892))
+
+
+
+
+
+# [3.60.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.59.3...@kong-ui-public/entities-shared@3.60.0) (2026-08-06)
+
+
+### Features
+
+* **entities-shared:** add OnboardingCard component ([#3658](https://github.com/Kong/public-ui-components/issues/3658)) ([9e8c160](https://github.com/Kong/public-ui-components/commit/9e8c1606d031536d2b154887186d33ff3f0df370))
+
+
+
+
+
 ## [3.59.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.59.2...@kong-ui-public/entities-shared@3.59.3) (2026-08-05)
 
 **Note:** Version bump only for package @kong-ui-public/entities-shared